move transitions.css to lib/transitions
