+++ app/lib/transitions/index.ts
@@ -1,0 +1,3 @@
+export * from './view-transition';
+import transitionsUrl from './transitions.css?url';
+export const transitionStyles = transitionsUrl;
