+++ app/tailwind.css
@@ -1,4 +1,3 @@
-@import "transitions.css";
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
